Add checkDimensions folder to an utility module in _lib     - Adjust remaining code to utilize this helper function     - Some of the code doesn't utilize it yet, change later
